PR review findings
