+++ src/main/java/bio/terra/workspace/service/workspace/flight/CreateSamResourceStep.java
@@ -6,6 +6,7 @@ import bio.terra.stairway.StepResult;
 import bio.terra.stairway.exception.RetryException;
 import bio.terra.workspace.service.iam.SamService;
 
+// TODO: PF-415 implement this
 public class CreateSamResourceStep implements Step {
 
   private final SamService samService;
