--- src/main/java/bio/terra/workspace/service/workspace/flight/FlightMapKey.java
@@ -1,11 +1,0 @@
-package bio.terra.workspace.service.workspace.flight;
-
-public interface FlightMapKey {
-  String getKey();
-
-  Class<?> getKlass();
-
-  //  default <T> T getFromMap(FlightMap flightMap) {
-  //    return flightMap.get(getKey(), getKlass());
-  //  }
-}
